requirements.txt
google-cloud-speech
spacy
firebase
pyrebase
mysql-connector-python
pandas
nltk
jsonlines